Move DAO classes in repositories.
