+++ src/main/java/repositories/UserDAO.java
@@ -1,4 +1,4 @@
-package services;
+package repositories;
 
 import models.User;
